[WEJBHTTP-150] Refactoring - introducing ServerHandlers class
